+++ game/server/EnvBeam.cpp
@@ -381,7 +381,7 @@ void CEnvBeam::Strike( void )
 	if ( pStart == NULL || pEnd == NULL )
 		return;
 
-	m_speed = clamp( m_speed, 0, MAX_BEAM_SCROLLSPEED );
+	m_speed = (int)clamp( m_speed, 0, MAX_BEAM_SCROLLSPEED );
 	
 	int pointStart = IsStaticPointEntity( pStart );
 	int pointEnd = IsStaticPointEntity( pEnd );
@@ -710,7 +710,7 @@ void CEnvBeam::BeamUpdateVars( void )
 
 	m_nNumBeamEnts = 2;
 
-	m_speed = clamp( m_speed, 0, MAX_BEAM_SCROLLSPEED );
+	m_speed = (int)clamp( m_speed, 0, MAX_BEAM_SCROLLSPEED );
 
 	// NOTE: If the end entity is the beam itself (and the start entity
 	// isn't *also* the beam itself, we've got problems. This is a problem
