+++ game/server/entitylist.cpp
@@ -6,6 +6,7 @@
 //=============================================================================//
 
 #include "cbase.h"
+
 #include "entitylist.h"
 #include "utlvector.h"
 #include "igamesystem.h"
@@ -143,7 +144,7 @@ public:
 	void Clear()
 	{
 		m_simThinkList.Purge();
-		for ( int i = 0; i < ARRAYSIZE(m_entinfoIndex); i++ )
+		for ( size_t i = 0; i < ARRAYSIZE(m_entinfoIndex); i++ )
 		{
 			m_entinfoIndex[i] = 0xFFFF;
 		}
