+++ game/server/fire.cpp
@@ -980,7 +980,7 @@ void CFire::Update( float simTime )
 	if ( m_flDamageTime <= gpGlobals->curtime )
 	{
 		m_flDamageTime = gpGlobals->curtime + fire_dmginterval.GetFloat();
-		outputDamage = (fire_dmgbase.GetFloat() + outputHeat * fire_dmgscale.GetFloat() * m_flDamageScale) * fire_dmginterval.GetFloat();
+		outputDamage = (int)((fire_dmgbase.GetFloat() + outputHeat * fire_dmgscale.GetFloat() * m_flDamageScale) * fire_dmginterval.GetFloat());
 		if ( outputDamage )
 		{
 			damage = true;
@@ -997,7 +997,7 @@ void CFire::Update( float simTime )
 		}
 		else if ( FClassnameIs( pOther, "env_fire" ) )
 		{
-			if ( fireCount < ARRAYSIZE(pFires) )
+			if ( fireCount < (int)ARRAYSIZE(pFires) )
 			{
 				pFires[fireCount] = (CFire *)pOther;
 				fireCount++;
