+++ game/server/hl2/cbasehelicopter.cpp
@@ -682,7 +682,7 @@ void CBaseHelicopter::UpdateEnemy()
 	// be sure and change my prevseen/lastseen timers.
 	if( m_lifeState == LIFE_ALIVE )
 	{
-		GetSenses()->Look( EnemySearchDistance() );
+		GetSenses()->Look( (int)EnemySearchDistance() );
 
 		GetEnemies()->RefreshMemories();
 		ChooseEnemy();
@@ -842,7 +842,7 @@ void CBaseHelicopter::UpdatePlayerDopplerShift( )
 			float velReceiver = DotProduct( pPlayer->GetAbsVelocity(), dir );
 			float velTransmitter = -DotProduct( GetAbsVelocity(), dir );
 			// speed of sound == 13049in/s
-			int iPitch = 100 * ((1 - velReceiver / 13049) / (1 + velTransmitter / 13049));
+			int iPitch = (int)(100 * ((1 - velReceiver / 13049) / (1 + velTransmitter / 13049)));
 #else
 			// This is a bogus doppler shift, but I like it better
 			float relV = DotProduct( GetAbsVelocity() - pPlayer->GetAbsVelocity(), dir );
