+++ game/server/hl2/grenade_bugbait.cpp
@@ -165,7 +165,7 @@ void CGrenadeBugBait::BugBaitTouch( CBaseEntity *pOther )
 		pSporeExplosion->SetLocalOrigin( GetAbsOrigin() );
 		pSporeExplosion->m_flSpawnRate			= 8.0f;
 		pSporeExplosion->m_flParticleLifetime	= 2.0f;
-		pSporeExplosion->SetRenderColor( 0.0f, 0.5f, 0.25f, 0.15f );
+		pSporeExplosion->SetRenderColor( 0, (int)0.5f, (int)0.25f, (int)0.15f );
 
 		pSporeExplosion->m_flStartSize = 32.0f;
 		pSporeExplosion->m_flEndSize = 64.0f;
