+++ game/server/hl2/item_ammo.cpp
@@ -32,7 +32,7 @@ int ITEM_GiveAmmo( CBasePlayer *pPlayer, float flCount, const char *pszAmmoName,
 	// Don't give out less than 1 of anything.
 	flCount = max( 1.0f, flCount );
 
-	return pPlayer->GiveAmmo( flCount, iAmmoType, bSuppressSound );
+	return pPlayer->GiveAmmo( (int)flCount, iAmmoType, bSuppressSound );
 }
 
 // ========================================================================
